% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_aug.R
\name{arrange_aug}
\alias{arrange_aug}
\title{Arrange aggregation matrix and apply different methods}
\usage{
arrange_aug(N, rep, dat1, dat2)
}
\arguments{
\item{N}{The number of aggregation times}

\item{rep}{The number of repeats for each aggregation time}

\item{dat1}{The list of one method from augmentation_matrix function}

\item{dat2}{The list of one method from augmentation_matrix function}
}
\description{
\code{arrange_aug} returns a data frame
}
\examples{
dat <- arrange_aug(50,10,aug_result1,aug_result2)
}
