% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.R
\name{augmentation_matrix}
\alias{augmentation_matrix}
\title{Create aggregation matrix and apply different methods}
\usage{
augmentation_matrix(N, M, n, seed, rep, method, bucket, reg_par, var)
}
\arguments{
\item{N}{Aggregation times}

\item{M}{The number of total sbm results}

\item{n}{The number of individuals}

\item{seed}{A vector of the seed you want to set for each aggregation time}

\item{rep}{The number of repeats for each aggregation time}

\item{method}{The method, kmeans, hclust or eigensplit}

\item{bucket}{The bucket where you save the sbm affinity matrix}

\item{reg_par}{The parameter value of regularization}

\item{var}{The percentage explaine the variation}
}
\description{
\code{create_affinity} returns an affinity matrix
}
\examples{
aug_result <- augmentation_matrix(20,2000,339,c(1:10),10,c("kmeans","hclust","eigensplit"),"jak2",0.1,0.9)
}
