% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanSBM.R
\name{cleanSBM}
\alias{cleanSBM}
\title{Run SBM and clean SBM results}
\usage{
cleanSBM(n, m, dat)
}
\arguments{
\item{n}{The number of times repeating running sbm}

\item{m}{The number of cores required, notice n/m should be integer}

\item{dat}{The data which satisfy the requirements of the formate of runnning sbm}
}
\value{
Lists of sbm results
}
\description{
\code{cleanSBM} returns lists of sbm results
}
\examples{
sbmresults <- cleanSBM(1000,10,pheno_dat)
}
