% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership_matrix.R
\name{membership}
\alias{membership}
\title{Build the Membership matrix}
\usage{
membership(dat, A)
}
\arguments{
\item{dat}{It is the result from partition_tree}

\item{A}{adjencency matrix}
}
\value{
A dataframe of membership matrix
}
\description{
\code{membership} returns a dataframe
}
\examples{
member <- membership(cluster_result,A)
}
